🧑‍💻- Move Backend (Golang version) outside Nx standard project structure.
