+++ backend/internal/app/main.go
@@ -4,9 +4,9 @@ import (
 	"context"
 	"fmt"
 	"github.com/gofiber/fiber/v3"
-	"github.com/solunion/way/backend/config"
-	"github.com/solunion/way/backend/http"
-	"github.com/solunion/way/backend/tenant"
+	"github.com/solunion/way/backend/internal/pkg/config"
+	"github.com/solunion/way/backend/internal/pkg/http"
+	"github.com/solunion/way/backend/internal/pkg/tenant"
 	"go.uber.org/fx"
 	"go.uber.org/fx/fxevent"
 	"go.uber.org/zap"
