--- go.work
@@ -1,10 +1,0 @@
-go 1.23.3
-
-use (
-	./apps/backend-new
-	./libs/backend-new/common
-	./libs/backend-new/config
-	./libs/backend-new/database
-	./libs/backend-new/http
-	./libs/backend-new/tenant
-)
