--- libs/backend-new/database/go.mod
@@ -1,5 +1,0 @@
-module github.com/solunion/way/backend/database
-
-go 1.23
-
-require github.com/google/uuid v1.6.0
