--- libs/backend-new/tenant/project.json
@@ -1,18 +1,0 @@
-{
-  "name": "github.com/solunion/way/backend/tenant",
-  "$schema": "../../../node_modules/nx/schemas/project-schema.json",
-  "projectType": "library",
-  "sourceRoot": "libs/backend-new/tenant",
-  "tags": [],
-  "targets": {
-    "test": {
-      "executor": "@nx-go/nx-go:test"
-    },
-    "lint": {
-      "executor": "@nx-go/nx-go:lint"
-    },
-    "tidy": {
-      "executor": "@nx-go/nx-go:tidy"
-    }
-  }
-}
